Renamed HUD to Attitude Indicator (AI)
