+++ DroidPlanner/src/org/droidplanner/fragments/TelemetryFragment.java
@@ -4,7 +4,7 @@ import org.droidplanner.DroidPlannerApp;
 import org.droidplanner.drone.Drone;
 import org.droidplanner.drone.DroneInterfaces.DroneEventsType;
 import org.droidplanner.drone.DroneInterfaces.OnDroneListener;
-import org.droidplanner.widgets.HUD.HUD;
+import org.droidplanner.widgets.AttitudeIndicator;
 
 import android.content.SharedPreferences;
 import android.os.Bundle;
@@ -19,7 +19,7 @@ import org.droidplanner.R;
 
 public class TelemetryFragment extends Fragment implements OnDroneListener{
 
-	private HUD hud;
+	private AttitudeIndicator attitudeIndicator;
 	private Drone drone;
 	private TextView roll;
 	private TextView yaw;
@@ -36,7 +36,7 @@ public class TelemetryFragment extends Fragment implements OnDroneListener{
 			Bundle savedInstanceState) {
 		View view = inflater.inflate(R.layout.fragment_telemetry, container,
 				false);
-		hud = (HUD) view.findViewById(R.id.hudView);
+		attitudeIndicator = (AttitudeIndicator) view.findViewById(R.id.aiView);
 
 		roll = (TextView) view.findViewById(R.id.rollValueText);
 		yaw = (TextView) view.findViewById(R.id.yawValueText);
@@ -95,7 +95,7 @@ public class TelemetryFragment extends Fragment implements OnDroneListener{
 			y = 360+y;
 		}
 
-		hud.setAttitude(r, p, y);
+		attitudeIndicator.setAttitude(r, p, y);
 
 		roll.setText(String.format("%3.0f\u00B0", r));
 		pitch.setText(String.format("%3.0f\u00B0", p));
